extracting the constructible test for transform_iterator to an independant test
